refactor apollo integration
