+++ lib/utils.js
@@ -82,26 +82,6 @@ export const singular = str => {
   return str.replace(/ies$/, 'y').replace(/s$/, '');
 };
 
-export const getBaseApiUrl = ({ internal } = {}) => {
-  if (process.browser) {
-    return '/api';
-  } else if (internal && process.env.INTERNAL_API_URL) {
-    return process.env.INTERNAL_API_URL;
-  } else {
-    return process.env.API_URL || 'https://api.opencollective.com';
-  }
-};
-
-/**
- * Returns the GraphQL api url for the appropriate api version and environment.
- * @param {string} version - api version. Defaults to v1.
- * @returns {string} GraphQL api url.
- */
-export const getGraphqlUrl = apiVersion => {
-  const apiKey = !process.browser ? process.env.API_KEY : null;
-  return `${getBaseApiUrl()}/graphql${apiVersion ? `/${apiVersion}` : ''}${apiKey ? `?api_key=${apiKey}` : ''}`;
-};
-
 export const capitalize = str => {
   if (typeof str !== 'string') {
     return '';
