+++ lib/withData.js
@@ -6,7 +6,7 @@ import PropTypes from 'prop-types';
 import { getDataFromTree } from '@apollo/client/react/ssr';
 import Head from 'next/head';
 
-import initClient from './initClient';
+import { initClient } from './apollo-client';
 
 // Gets the display name of a JSX component for dev tools
 function getComponentDisplayName(Component) {
@@ -15,55 +15,21 @@ function getComponentDisplayName(Component) {
 
 const withData = ComposedComponent => {
   return class WithData extends React.Component {
-    static async getInitialProps(ctx) {
-      // Initial serverState with apollo (empty)
-      let serverState = {
-        apollo: {
-          data: {},
-        },
-      };
-      const { Component, router } = ctx;
-
-      const options = {
-        headers: ctx.req ? ctx.req.headers : {},
-      };
+    static async getInitialProps(context) {
+      const { Component } = context;
 
       // Evaluate the composed component's getInitialProps()
       let composedInitialProps = {};
       if (ComposedComponent.getInitialProps) {
-        composedInitialProps = await ComposedComponent.getInitialProps(ctx);
+        composedInitialProps = await ComposedComponent.getInitialProps(context);
       }
 
       // Run all GraphQL queries in the component tree
       // and extract the resulting data
-      const apollo = initClient(undefined);
+      const client = initClient();
       try {
-        // create the url prop which is passed to every page
-        const url = {
-          query: ctx.query,
-          asPath: ctx.asPath,
-          pathname: ctx.pathname,
-        };
-
         // Run all GraphQL queries
-        await getDataFromTree(
-          <ComposedComponent
-            ctx={ctx}
-            url={url}
-            client={apollo}
-            Component={Component}
-            router={router}
-            {...composedInitialProps}
-          />,
-          {
-            router: {
-              asPath: ctx.asPath,
-              pathname: ctx.pathname,
-              query: ctx.query,
-            },
-            client: apollo,
-          },
-        );
+        await getDataFromTree(<ComposedComponent Component={Component} {...composedInitialProps} />, { client });
       } catch (error) {
         // Prevent Apollo Client GraphQL errors from crashing SSR.
         // Handle them in components via the data.error prop:
@@ -79,14 +45,13 @@ const withData = ComposedComponent => {
       }
 
       // Extract query data from the Apollo store
-      serverState = {
+      const serverState = {
         apollo: {
-          data: apollo.cache.extract(),
+          data: client.cache.extract(),
         },
       };
 
       return {
-        options,
         serverState,
         ...composedInitialProps,
       };
@@ -110,11 +75,11 @@ const withData = ComposedComponent => {
     constructor(props) {
       super(props);
       const { serverState } = this.props;
-      this.apollo = initClient(serverState.apollo.data);
+      this.client = initClient({ initialState: serverState.apollo.data });
     }
 
     render() {
-      return <ComposedComponent {...this.props} client={this.apollo} />;
+      return <ComposedComponent {...this.props} client={this.client} />;
     }
   };
 };
