+++ pages/unsubscribeEmail.js
@@ -3,8 +3,6 @@ import PropTypes from 'prop-types';
 import { Email } from '@styled-icons/material/Email';
 import { FormattedMessage } from 'react-intl';
 
-import { getBaseApiUrl } from '../lib/utils';
-
 import Container from '../components/Container';
 import { Box } from '../components/Grid';
 import MessageBox from '../components/MessageBox';
@@ -40,9 +38,7 @@ class UnsubscribeEmail extends React.Component {
   async componentDidMount() {
     let state, errorMessage, response;
     await fetch(
-      `${getBaseApiUrl()}/services/email/unsubscribe/${this.props.email}/${this.props.slug}/${this.props.type}/${
-        this.props.token
-      }`,
+      `/api/services/email/unsubscribe/${this.props.email}/${this.props.slug}/${this.props.type}/${this.props.token}`,
     ).then(res => {
       response = res.json();
     });
