+++ test/providers.js
@@ -4,7 +4,7 @@ import { get } from 'lodash';
 import { IntlProvider } from 'react-intl';
 import { ThemeProvider } from 'styled-components';
 
-import initClient from '../lib/initClient';
+import { initClient } from '../lib/apollo-client';
 import theme from '../lib/theme';
 import * as Intl from '../server/intl';
